Initial Commit
Beginnings of the Lexer. There is no build system yet but this will
come soon. For the moment there are only two files to be compiled.

Pine.l is the lexer (hopefully almost complete) for pine. The parser
is the most important part of this poject and there is a sketch for
the language it will accept in `pine.language`.