Create Symbol Class
Create the symbol class, also create the unit tests for the new class. This is
part of the new refactor of the codebase. We will be using collections of
these Symbol classes in Expansion classes to represent the expansions of a
Symbol and these non-terminal symbols will then be grouped together in a
Grammar class.

* new class: Symbol
* new unit tests
* new dependancy: nose (for running the tests)