Enable Python 2.7 Use
I don't really have much intention of ensuring that this module can be used from
older versions of python, however if it does't cost too much then it would be a
nice-to-have feature. This commit makes the current tree work with Python 2.7.
I don't forsee support for any versions before the backporting of Python 3
features.