Add Mustache Render Context
This context uses the mustache templating system to write out files. Language
defnitions are then made in `.mustache` files and a simple wrapper class is used
to render them.

This is still a bit messy as there is quite a bit of logic in the views (both
callback and mustache). It would be better if some of this computation could
be moved into the grammar class.