Add TODO Items
Add a file containing TODO items for the project. The plan is to put thoughts
in here and then move them to issues when I feel like working on them. This
means that short-lived TODOs don't need an issue. It also means that others can
see what I plan for the project and chip in if they feel like it.