Update Command Line Application
Command line useage is now possible. The app now takes two command line
options to specify the language to use and the place to output the results of
the grammar transformation. Option overrides from the command line are also now
passed through to the output context allowing them to affect how the whole thing works.